backend/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
caer==2.0.8
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
convertapi==2.0.0
Django==4.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.15.2
docx2pdf==0.1.8
gevent==24.2.1
greenlet==3.1.1
idna==3.7
lxml==5.3.0
mypy==1.11.2
mypy-extensions==1.0.0
numpy==1.26.4
opencv-contrib-python-headless==4.10.0.84
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
psutil==7.0.0
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMuPDF==1.24.13
pypandoc==1.14
PyPDF2==3.0.1
python-decouple==3.8
python-docx==1.1.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.3
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
websocket==0.2.1
websockets==13.1
zope.event==5.0
zope.interface==7.0.3
dj-database-url
python-dotenv
drf_spectacular
drf_spectacular_sidecar
pandas
openpyxl
getmac
django-simple-history
psycopg2-binary
transliterate